committing Nick Zitzmann's patch with additional project changes in order for the prefpane to actually load in 64-bit mode under System Preferences, oddly we were linking to SystemUIPlugin.framework unnecessarily which was causing a GC mismatch and resulting it in relaunching in 32-bit mode

+++ Common/Source/GrowlPathUtilities.m
@@ -35,7 +35,7 @@ restart:;
 	UInt32 oldestProcessLaunchDate = UINT_MAX;
 
 	while ((err = GetNextProcess(&psn)) == noErr) {
-		struct ProcessInfoRec info = { .processInfoLength = sizeof(struct ProcessInfoRec) };
+		struct ProcessInfoRec info = { .processInfoLength = (UInt32)sizeof(struct ProcessInfoRec) };
 		err = GetProcessInformation(&psn, &info);
 		if (err == noErr) {
 			//Compare the launch dates first, since it's cheaper than comparing bundle IDs.
