+++ Plugins/Displays/SMS/GrowlSMSPrefs.m
@@ -80,8 +80,8 @@
 	UInt32 passwordLength;
 	OSStatus status;
 	status = SecKeychainFindGenericPassword( NULL,
-											 strlen(keychainServiceName), keychainServiceName,
-											 strlen(keychainAccountName), keychainAccountName,
+											 (UInt32)strlen(keychainServiceName), keychainServiceName,
+											 (UInt32)strlen(keychainAccountName), keychainAccountName,
 											 &passwordLength, (void **)&password, NULL );
 
 	NSString *passwordString;
@@ -100,28 +100,28 @@
 
 - (void) setAccountPassword:(NSString *)value {
 	const char *password = value ? [value UTF8String] : "";
-	unsigned length = strlen(password);
+	UInt32 length = (UInt32)strlen(password);
 	OSStatus status;
 	SecKeychainItemRef itemRef = nil;
 	status = SecKeychainFindGenericPassword( NULL,
-											 strlen(keychainServiceName), keychainServiceName,
-											 strlen(keychainAccountName), keychainAccountName,
+											 (UInt32)strlen(keychainServiceName), keychainServiceName,
+											 (UInt32)strlen(keychainAccountName), keychainAccountName,
 											 NULL, NULL, &itemRef );
 	if (status == errSecItemNotFound) {
 		// add new item
 		status = SecKeychainAddGenericPassword( NULL,
-												strlen(keychainServiceName), keychainServiceName,
-												strlen(keychainAccountName), keychainAccountName,
+												(UInt32)strlen(keychainServiceName), keychainServiceName,
+												(UInt32)strlen(keychainAccountName), keychainAccountName,
 												length, password, NULL );
 		if (status)
 			NSLog(@"Failed to add SMS Account password to keychain.");
 	} else {
 		// change existing password
 		SecKeychainAttribute attrs[] = {
-			{ kSecAccountItemAttr, strlen(keychainAccountName), (char *)keychainAccountName },
-			{ kSecServiceItemAttr, strlen(keychainServiceName), (char *)keychainServiceName }
+			{ kSecAccountItemAttr, (UInt32)strlen(keychainAccountName), (char *)keychainAccountName },
+			{ kSecServiceItemAttr, (UInt32)strlen(keychainServiceName), (char *)keychainServiceName }
 		};
-		const SecKeychainAttributeList attributes = { sizeof(attrs) / sizeof(attrs[0]), attrs };
+		const SecKeychainAttributeList attributes = { (UInt32)sizeof(attrs) / (UInt32)sizeof(attrs[0]), attrs };
 		status = SecKeychainItemModifyAttributesAndData( itemRef,		// the item reference
 														 &attributes,	// no change to attributes
 														 length,		// length of password
